adds ResourceIntegerValue class that gets resource dimensions changes resource dimensions from doubles to integers

+++ modeller-fedora/src/main/java/org/blume/modeller/templates/CanvasScope.java
@@ -63,8 +63,8 @@ public class CanvasScope extends Scope {
         String resourceURI;
         String listURI;
         String canvasLabel;
-        double canvasHeight;
-        double canvasWidth;
+        int canvasHeight;
+        int canvasWidth;
     }
 
     public static class Prefix {
